+++ rusty_box/src/cpu/softfloat3e/i32_to_extf80.rs
@@ -18,7 +18,7 @@ pub fn i32_to_extf80(a: i32) -> floatx80 {
     let abs_a = if sign { (-(a as i64)) as u32 } else { a as u32 };
     let shift_dist = count_leading_zeros32(abs_a);
     let ui_z64 = pack_to_extf80_sign_exp(sign, (0x401E - shift_dist as i32) as u16);
-    let ui_z0 = (((abs_a as u64) << shift_dist) as u64) << 32;
+    let ui_z0 = ((abs_a as u64) << shift_dist) << 32;
 
     floatx80 {
         signif: ui_z0,
